Add new value stack
